♻️ Move `analytics` to TS
