+++ src/utils/analytics/__tests__/stubs.ts
